Backend/Kineayuda-main/requirements.txt
anyio==4.11.0
asgiref==3.9.1
CacheControl==0.14.3
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==46.0.3
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
filelock==3.19.1
firebase_admin==7.1.0
fsspec==2025.9.0
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==3.26.1
mpmath==1.3.0
msgpack==1.1.2
networkx==3.5
numpy==2.3.3
packaging==25.0
pillow==12.0.0
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transbank-sdk==6.1.0
transformers==4.57.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
pysentimiento==0.7.2
finiteautomata==0.1.5

